move nav to its own folder
