+++ src/components/loader.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="flex justify-center items-center opacity-50">
-    <img src="../assets/loader.gif" />
+    <img src="@/assets/loader.gif" />
   </div>
 </template>
